Score tracking overhaul
I am happy to say that it was just tweaking the code with one addition
of another array of "score" to patch this one around and not a complete
rewrite of code. This requires any previously build games to be
regenerated through the generator but overall reliability seems greatly
improved.